pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "colaber"
version = "0.1.0"
description = "Run local Python scripts on Google Colab GPU/TPU runtimes"
license = "MIT"
authors = [{ name = "Clark Miyamoto" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "websockets>=12.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=0.7.0",
    "pathspec>=0.11",
    "python-dotenv>=1.0",
]

[project.scripts]
colaber = "colaber.cli:main"

[project.optional-dependencies]
dev = ["pytest", "ruff"]
